fixat peters hjärna

+++ models.py
@@ -18,7 +18,7 @@ def FCN(no_inputs, no_outputs, no_layers, no_nodes):
     Returns:
         fully-connected neural network as tensorflow.keras.Model object
     """
-    inputs = Input(shape=(no_inputs,), dtype='int32')
+    inputs = Input(shape=(no_inputs,), dtype='float32')
     x = Dense(no_nodes, activation='relu')(inputs)
     for i in range(no_layers-2):
         x = Dense(no_nodes, activation='relu')(x)
